Move Notifications component to its own directory
